Conflicts:
	lib/kill.js
	src/kill.coffee

+++ src/kill.coffee
@@ -6,21 +6,20 @@ TIMEOUT  = 60000
 
 killTree = (pid, signal, timeout, done) ->
   psTree pid, (err, children) ->
+    return done(err) if err
+
     pids = children.map (c) -> c.PID
     kill = (pid, cb) -> killProcess(pid, signal, timeout, cb)
 
     async.series [
-      (cb) -> async.each pids, kill, cb
+      (cb) -> async.each(pids, kill, cb)
       (cb) -> kill(pid, cb)
     ], done
 
 killProcess = (pid, signal, timeout, cb) ->
-  try
-    process.kill(pid, signal)
-  catch e
-    console.log "failed to kill #{pid}:", e
-    cb null
+  return cb() unless common.checkProcess(pid)
 
+  process.kill(pid, signal)
   start = Date.now()
   test  = ->
     running  = isRunning(pid)
